pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "imagetrust"
version = "0.1.0"
description = "A Forensic Application for Identifying AI-Generated and Digitally Manipulated Images"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.10"
keywords = [
    "ai-detection",
    "image-forensics",
    "deepfake",
    "c2pa",
    "machine-learning",
    "computer-vision",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "timm>=0.9.0",
    "transformers>=4.35.0",
    "exifread>=3.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "click>=8.1.0",
    "rich>=13.0.0",
    "streamlit>=1.30.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "tqdm>=4.66.0",
    "reportlab>=4.0.0",
    "jinja2>=3.1.0",
    "matplotlib>=3.8.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.6.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.24.0",
]
explainability = [
    "pytorch-grad-cam>=1.4.8",
    "captum>=0.6.0",
    "shap>=0.43.0",
]
c2pa = [
    "c2pa-python>=0.4.0",
    "python-xmp-toolkit>=2.0.1",
]
baselines = [
    "xgboost>=2.0.0",              # For XGBoost classical baseline
    "scikit-image>=0.22.0",        # For LBP feature extraction
]
desktop = [
    "PySide6>=6.6.0",              # Qt GUI for desktop app
    "pyinstaller>=6.0.0",          # For building .exe
]
clip = [
    "ftfy>=6.1.0",                 # Required by CLIP
    "regex>=2023.0.0",             # Required by CLIP
    # Note: CLIP installed via: pip install git+https://github.com/openai/CLIP.git
]
all = [
    "imagetrust[dev,docs,explainability,c2pa,baselines,desktop]",
]

[project.scripts]
imagetrust = "imagetrust.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/imagetrust"
Documentation = "https://imagetrust.readthedocs.io"
Repository = "https://github.com/yourusername/imagetrust"
Issues = "https://github.com/yourusername/imagetrust/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
imagetrust = ["*.yaml", "*.json", "templates/*", "static/*"]

# Black configuration
[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["imagetrust"]

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

# MyPy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "tests",
    "docs",
]

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "gpu: marks tests that require GPU",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/imagetrust"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
